Reorganized to the 'standard' project layout
