--- main.go
@@ -1,13 +1,0 @@
-package main
-
-import (
-	"os"
-
-	"github.com/engineyard/scaley/cmd"
-)
-
-func main() {
-	if cmd.Execute() != nil {
-		os.Exit(1)
-	}
-}
