+++ pkg/common/locations_test.go
