+++ pkg/notifier/model.go
