+++ scripts/genver
